+++ NEWS.md
@@ -1,5 +1,7 @@
 # gtsummary (development version)
 
+* The `with_gtsummary_theme()` has been updated to no longer print theme names when the applied, nor when the original theme is re-applied. (#2031)
+
 * Updated the `theme_gtsummary_journal("jama")` theme to apply changes to `tbl_svysummary()`. (#1964; @vjcatharine)
 
 * Removed `global_pvalue_fun.tidycrr()` as already been migrated to the {tidycmprsk} package. (#1997; @jwoolfolk)
@@ -7,6 +9,8 @@
 * Bug fix in `add_ci.tbl_svysummary()` for factor variables where order was alphabetical instead of the factor levels. (#2036)
 
 * Addressing encoding issue where `sort()` and `dplyr::arrange()` sorted differently, and the order of the `by` levels was inconsistent in the resulting table. (#2038)
+
+* Added S3 methods `add_overall.tbl_hierarchical()` and `add_overall.tbl_hierarchical_count()`.
 
 # gtsummary 2.0.3